+++ packages/auth-zkp-iden3/src/types/index.ts
@@ -1,2 +1,1 @@
 export * from './auth-zkp-types'
-export * from './contracts'