+++ packages/shared-zkp-iden3/src/errors/index.ts
@@ -1,0 +1,7 @@
+import * as runtimeErrors from './runtime-errors'
+
+export const errors = {
+  ...runtimeErrors,
+}
+
+export * from './runtime-errors'