+++ packages/shared-zkp-iden3/src/helpers/state-v2-helpers.ts
@@ -1,21 +1,23 @@
+import type { Provider } from '@ethersproject/providers'
+import type { RawProvider } from '@rarimo/provider'
 import { providers } from 'ethers'
 
 import { StateV2__factory } from '@/types'
 
 export const getGISTProof = async ({
   rpcUrl,
+  rawProvider,
   contractAddress,
   userId,
 }: {
-  rpcUrl: string
+  rpcUrl?: string
+  rawProvider?: RawProvider
   contractAddress: string
   userId: string
 }) => {
-  const rawProvider = new providers.JsonRpcProvider(rpcUrl, 'any')
-
   const contractInstance = StateV2__factory.connect(
     contractAddress,
-    rawProvider,
+    (rawProvider || new providers.JsonRpcProvider(rpcUrl, 'any')) as Provider,
   )
 
   return contractInstance.getGISTProof(userId)