+++ packages/zkp-gen-iden3/src/types/zkp-gen-types.ts
@@ -1,14 +1,15 @@
 import type { Hash, Siblings } from '@iden3/js-merkletree'
 import type { VerifiableCredentials } from '@rarimo/auth-zkp-iden3'
 import type { Identity } from '@rarimo/identity-gen-iden3'
+import type { RawProvider } from '@rarimo/provider'
 
-import type { ZKP_OPERATORS } from '@/enums'
+import type { ZkpOperators } from '@/enums'
 
 export type QueryVariableNameAbstract = { [key: string]: number }
 
 export type ZkpGenQuery<T extends QueryVariableNameAbstract> = {
   variableName: keyof T
-  operator: ZKP_OPERATORS
+  operator: ZkpOperators
   value: string[]
 }
 
@@ -48,4 +49,14 @@ export type Schema = {
       jsonSchema: string
     }
   }
+}
+
+export type Config = {
+  RPC_URL?: string
+  RAW_PROVIDER?: RawProvider
+  ISSUER_API_URL: string
+  STATE_V2_ADDRESS: string
+  CIRCUIT_WASM_URL: string
+  CIRCUIT_FINAL_KEY_URL: string
+  CLAIM_PROOF_SIBLINGS_COUNT: number
 }