+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `root` - `@rarimo/identity-gen-iden3` package
 - `root` - `@rarimo/auth-zkp-iden3` package
 - `root` - `@rarimo/zkp-gen-iden3` package
+- `root` - `@rarimo/shared-zkp-iden3` package
 - `@rarimo/shared` - export createApi method
 
 ## [2.0.0-rc.2] - 2023-06-13